implement surfaceField and volumeField as geometricField
