+++ include/NeoFOAM/finiteVolume/cellCentred.hpp
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: MIT
 // SPDX-FileCopyrightText: 2024 NeoFOAM authors
 
+#include "cellCentred/fields/geometricField.hpp"
 #include "cellCentred/fields/surfaceField.hpp"
 #include "cellCentred/fields/volumeField.hpp"
 #include "cellCentred/fields/boundary/boundaryField.hpp"
