+++ include/NeoFOAM/finiteVolume/cellCentred/fields/volumeField.hpp
@@ -23,103 +23,11 @@ namespace NeoFOAM::finiteVolume::cellCentred
  * @tparam ValueType The type of the field values.
  */
 template<typename ValueType>
-class VolumeField
+class VolumeField final : public GeometricField<ValueType, BoundaryField<ValueType>>
 {
 public:
 
-    /**
-     * @brief Constructor for VolumeField.
-     *
-     * @param exec The executor for parallel execution.
-     * @param mesh The unstructured mesh.
-     * @param boundaryConditions The boundary conditions for the field.
-     */
-    VolumeField(const Executor& exec)
-        : exec_(exec), mesh_(nullptr), field_(exec), boundaryConditions_()
-    {}
-
-    /**
-     * @brief Constructor for VolumeField.
-     *
-     * @param exec The executor for parallel execution.
-     * @param mesh The unstructured mesh.
-     * @param boundaryConditions The boundary conditions for the field.
-     */
-    VolumeField(
-        const Executor& exec,
-        std::shared_ptr<const UnstructuredMesh> mesh,
-        std::vector<std::unique_ptr<BoundaryField<ValueType>>>&& boundaryConditions
-    )
-        : exec_(exec), mesh_(mesh),
-          field_(exec, mesh->nCells(), mesh->nBoundaryFaces(), mesh->nBoundaries()),
-          boundaryConditions_(std::move(boundaryConditions))
-    {}
-
-    /**
-     * @brief Corrects the boundary conditions of the field.
-     *
-     * This method applies the correctBoundaryConditions() method to each boundary condition
-     * in the field.
-     */
-    void correctBoundaryConditions()
-    {
-        for (auto& boundaryCondition : boundaryConditions_)
-        {
-            boundaryCondition->correctBoundaryConditions(
-                field_.boundaryField(), field_.internalField()
-            );
-        }
-    }
-
-    /**
-     * @brief Returns a const reference to the internal field.
-     *
-     * @return A const reference to the internal field.
-     */
-    const Field<ValueType>& internalField() const { return field_.internalField(); }
-
-    /**
-     * @brief Returns a reference to the internal field.
-     *
-     * @return A reference to the internal field.
-     */
-    Field<ValueType>& internalField() { return field_.internalField(); }
-
-    /**
-     * @brief Returns a const reference to the boundary field.
-     *
-     * @return A const reference to the boundary field.
-     */
-    const BoundaryFields<ValueType>& boundaryField() const { return field_.boundaryField(); }
-
-    /**
-     * @brief Returns a reference to the boundary field.
-     *
-     * @return A reference to the boundary field.
-     */
-    BoundaryFields<ValueType>& boundaryField() { return field_.boundaryField(); }
-
-    /**
-     * @brief Returns a const reference to the executor.
-     *
-     * @return A const reference to the executor.
-     */
-    const Executor& exec() const { return exec_; }
-
-    /**
-     * @brief Returns a const reference to the unstructured mesh object.
-     *
-     * @return The const reference to the unstructured mesh object.
-     */
-    std::shared_ptr<const UnstructuredMesh> mesh() const { return mesh_; }
-
-private:
-
-    Executor exec_;                                /**< The executor for parallel execution. */
-    std::shared_ptr<const UnstructuredMesh> mesh_; /**< The unstructured mesh. */
-    DomainField<ValueType> field_;                 /**< The domain field. */
-    std::vector<std::unique_ptr<BoundaryField<ValueType>>>
-        boundaryConditions_; /**< The boundary conditions for the field. */
+    using GeometricField<ValueType, BoundaryField<ValueType>>::GeometricField;
 };
 
 } // namespace NeoFOAM
