+++ include/NeoFOAM/mesh/unstructured.hpp
@@ -1,0 +1,5 @@
+// SPDX-License-Identifier: MIT
+// SPDX-FileCopyrightText: 2024 NeoFOAM authors
+
+#include "NeoFOAM/mesh/unstructured/boundaryMesh.hpp"
+#include "NeoFOAM/mesh/unstructured/unstructuredMesh.hpp"
